1. Renamed render.py to server.py 2. Added rcontroller (executable) to simplify launching (currently the same as __main__.py)
--server.py--
Changed a bunch of stuff so this can be imported as a module. Created Server class as module master and moved all the server-related stuff in there. Created Config class to hold config vars as class attributes. Added methods to export the current state of all Job instances to a file and to read that file on startup and restore the server's queue to its previous state.

--gui.py--
Config class now stores config vars as class attributes instead of instance attributes for consistency with server.py and so things work if file is imported as a module and MasterWin is instantiated. Now MasterWin passes itself as a param to classes it instantiates that may need to call methods in the parent MasterWin instance.